+++ packages/login-panel/src/actions.js
@@ -7,7 +7,8 @@ export const submitLogin = async (
   url,
   data,
   authenticityToken,
-  trackingData
+  trackingData,
+  buttonTrack
 ) => {
   const config = {
     headers: {
@@ -22,7 +23,7 @@ export const submitLogin = async (
     .then(response => {
       const { data, status } = response;
 
-      track('signin/submit', trackingData);
+      track(`${buttonTrack ? buttonTrack : 'signin'}/submit`, trackingData);
 
       return { status, data };
     })
@@ -30,7 +31,11 @@ export const submitLogin = async (
       const { data, status } = error.response;
 
       // TODO: replace this with a proper error logging (e.g airbrake)
-      track('signin/submit_error', trackingData);
+
+      track(
+        `${buttonTrack ? buttonTrack : 'signin'}/submit_error`,
+        trackingData
+      );
 
       if (status !== 401) {
         return {